refactor: move CSS files to styles directory
